began refactor - created basic file structure - copied class routes and handlers
